+++ examples/scripts/BPX_spm.py
@@ -46,18 +46,18 @@ cost = pybop.SumSquaredError(problem)
 optim = pybop.CMAES(cost, max_iterations=40)
 
 # Run the optimisation
-x, final_cost = optim.run()
+results = optim.run()
 print("True parameters:", parameters.true_value())
-print("Estimated parameters:", x)
+print("Estimated parameters:", results.x)
 
 # Plot the timeseries output
-pybop.quick_plot(problem, problem_inputs=x, title="Optimised Comparison")
+pybop.plot.quick(problem, problem_inputs=results.x, title="Optimised Comparison")
 
 # Plot convergence
-pybop.plot_convergence(optim)
+pybop.plot.convergence(optim)
 
 # Plot the parameter traces
-pybop.plot_parameters(optim)
+pybop.plot.parameters(optim)
 
-# Plot the cost landscape with optimisation path and updated bounds
-pybop.plot2d(optim, steps=5)
+# Plot the cost landscape with optimisation path
+pybop.plot.surface(optim)