+++ examples/scripts/ecm_tau_constraints.py
@@ -178,18 +178,18 @@ optim = pybop.XNES(
     max_iterations=125,
 )
 
-x, final_cost = optim.run()
-print("Estimated parameters:", x)
+results = optim.run()
+print("Estimated parameters:", results.x)
 
 
 # Plot the time series
-pybop.plot_dataset(dataset)
+pybop.plot.dataset(dataset)
 
 # Plot the timeseries output
-pybop.quick_plot(problem, problem_inputs=x, title="Optimised Comparison")
+pybop.plot.quick(problem, problem_inputs=results.x, title="Optimised Comparison")
 
 # Plot convergence
-pybop.plot_convergence(optim)
+pybop.plot.convergence(optim)
 
 # Plot the parameter traces
-pybop.plot_parameters(optim)
+pybop.plot.parameters(optim)